+++ resources/views/components/layout.blade.php
@@ -1,0 +1,18 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>{{ $title }}</title>
+</head>
+<body>
+    <nav>
+        <x-navlink href="/" :active="request()->is('home')">Home</x-navlink>
+        <x-navlink href="{{ route('articles.index') }}" :active="request()->is('articles.index')">Articles</x-navlink>
+        <x-navlink href="/dashboard" :active="request()->is('dashboard')">Dashboard</x-navlink>
+    </nav>
+    {{ $slot }}
+</body>
+</html>