+++ resources/views/home.blade.php
@@ -1,2 +1,4 @@
-<h1>Welcome!</h1>
-<p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Amet atque eveniet excepturi fugit, incidunt iste optio quibusdam tempore. Beatae dolor fuga iure libero, maiores modi necessitatibus nulla sapiente velit voluptas.</p>
+<x-layout title="Home">
+    <h1>Welcome to the home page</h1>
+    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Aspernatur atque debitis dicta dolorum, eius et eum eveniet, impedit molestias nostrum numquam, obcaecati perferendis quae ullam vitae. Aliquam excepturi obcaecati quas.</p>
+</x-layout>