+++ routes/web.php
@@ -1,15 +1,17 @@
 <?php
 
+use App\Http\Controllers\ArticleController;
+use App\Http\Controllers\HomeController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
+Route::get('/welcome', function () {
     return view('welcome');
 });
 
-Route::get('/home', function () {
-    return view('home');
-});
+Route::get('/',[HomeController::class, 'index']);
+Route::resource('articles', ArticleController::class);
+Route::delete('articles/{article}', [ArticleController::class, 'destroy'])->name('articles.destroy');
 
 Route::get('/dashboard', function () {
     return view('dashboard');
@@ -20,5 +22,7 @@ Route::middleware('auth')->group(function () {
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
+
+
 
 require __DIR__.'/auth.php';