+++ _changelog/changelog.md
@@ -1,8 +1,19 @@
 # Changelog
 
+### 16/10/2024
+* Delete functie toegevoegd
+* Article create pagina toegevoegd
+
+### 15/10/2024
+* Article controller toegevoegd
+* Articles pagina toegevoegd
+* Migrations toegevoegd (Articles staan nu in database)
+* Articles uit de database staan op articles.index
+
 ### 14/10/2024
 * Home pagina toegevoegd
 * Home controller toegevoegd
+* Components toegevoegd
 
 ### 09/10/2024
 * Project aangemaakt